--- src/components/Card1/Card1.scss
@@ -1,89 +1,0 @@
-@import "../../styles/ColorSchema.scss";
-
-.container {
-  position: absolute;
-  left: 0px;
-  top: 0px;
-  width: 1024px;
-  height: 768px;
-  z-index: 8;
-}
-.Button_text {
-  font-size: 12px;
-  font-family: "Open Sans";
-  color: $white;
-  line-height: 1.402;
-  text-align: center;
-  position: absolute;
-  padding: 30px;
-  margin-top: -17px;
-  height: 9px;
-  z-index: 7;
-}
-.button {
-  border-width: 0.22px;
-  border-radius: 21px;
-  border-color: $black;
-  border-style: solid;
-  background-color: $blue;
-  position: absolute;
-  left: 570px;
-  top: 537px;
-  width: 134px;
-  height: 41px;
-  z-index: 6;
-}
-.text {
-  font-size: 12px;
-  font-family: "Open Sans";
-  color: $black;
-  line-height: 1.402;
-  text-align: left;
-  position: absolute;
-  padding: 30px;
-  z-index: 5;
-  width: 200px;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-.Title__text {
-  font-size: 19px;
-  font-family: "Open Sans";
-  color: $black;
-  line-height: 1.402;
-  text-align: center;
-  position: absolute;
-  //   left: 486.445px;
-  //   top: 225.558px;
-  padding: 20px;
-  width: 24px;
-  height: 9px;
-  z-index: 4;
-}
-.title_rectangle {
-  border-width: 0.22px;
-  border-color: $black;
-  border-style: solid;
-  border-radius: 21px;
-  background-color: $blue;
-  position: absolute;
-  left: 435px;
-  top: 191px;
-  width: 163px;
-  height: 50px;
-  z-index: 3;
-}
-.card {
-  border-width: 0.22px;
-  border-color: $black;
-  border-style: solid;
-  border-radius: 10px;
-  background-color: $white;
-  position: absolute;
-  left: 435px;
-  top: 216px;
-  width: 269px;
-  height: 341px;
-  z-index: 2;
-}
