+++ src/components/index.js
@@ -7,6 +7,6 @@ export { default as Grid } from "./Grid";
 export { default as Media } from "./Media";
 export { default as Switch } from "./Switch";
 export { default as ZapBar } from "./ZapBar";
-export { default as Card1 } from "./Card1";
 export { default as Colors } from "./Colors";
 export { default as ZapButton } from "./ZapButton";
+export { default as Card } from "./Card";
